Move background tests to tests/ folder
