+++ test/background/providers/rss/tests/parse-feeds.test.js
@@ -2,7 +2,7 @@
 
 import {test} from 'ava';
 
-import FeedParser from '../lib/parse-feed';
+import FeedParser from '../../../../../src/background/providers/rss/lib/parse-feed';
 
 import {
     generateAtomFeed,
